House Price Prediction
This dataset focuses on housing prices in Ames, Iowa, and is intended for predictive modeling of property values. It contains detailed information about various attributes of residential properties, including physical characteristics, quality and condition, and their specific locations within the city. The dataset is organized into several files: train.csv for model training, test.csv for evaluation, and sample_submission.csv which includes a benchmark prediction based on a linear regression model. The dataset features columns that cover fundamental details such as lot area and bedroom count, as well as more specific attributes like basement quality and garage type. The goal is to predict the SalePrice, which represents the sale price of each property in dollars. The accompanying data description file offers comprehensive explanations of each column, aiding in the understanding and preprocessing necessary for predictive analysis.